requirements.txt
streamlit==1.45.1
tensorflow==2.13.0
pillow==11.2.0
numpy==2.3.0
